+++ src/components/Channel/BasicInfo/Subscribe.tsx
@@ -1,6 +1,6 @@
 import { LENSHUB_PROXY_ABI } from '@abis/LensHubProxy'
 import { useMutation } from '@apollo/client'
-import { Button } from '@components/ui/Button'
+import { Button } from '@components/UIElements/Button'
 import { LENSHUB_PROXY_ADDRESS } from '@utils/constants'
 import omitKey from '@utils/functions/omitKey'
 import { CREATE_FOLLOW_TYPED_DATA } from '@utils/gql/queries'
