+++ src/components/Channel/Upload/Details.tsx
@@ -1,10 +1,10 @@
 import { LENSHUB_PROXY_ABI } from '@abis/LensHubProxy'
 import { useMutation } from '@apollo/client'
 import { WebBundlr } from '@bundlr-network/client'
-import VideoPlayer from '@components/common/VideoPlayer'
-import { Button } from '@components/ui/Button'
-import ChooseImage from '@components/ui/ChooseImage'
-import { Input } from '@components/ui/Input'
+import VideoPlayer from '@components/Common/VideoPlayer'
+import { Button } from '@components/UIElements/Button'
+import ChooseImage from '@components/UIElements/ChooseImage'
+import { Input } from '@components/UIElements/Input'
 import useAppStore from '@lib/store'
 import {
   BUNDLR_CURRENCY,
