+++ src/components/Channel/index.tsx
@@ -1,6 +1,6 @@
 import { useQuery } from '@apollo/client'
-import Layout from '@components/common/Layout'
-import MetaTags from '@components/common/MetaTags'
+import Layout from '@components/Common/Layout'
+import MetaTags from '@components/Common/MetaTags'
 import ChannelShimmer from '@components/Shimmers/ChannelShimmer'
 import { PROFILE_QUERY } from '@utils/gql/queries'
 import dynamic from 'next/dynamic'
