+++ src/components/Home/Feed.tsx
@@ -1,7 +1,7 @@
 import { useQuery } from '@apollo/client'
 import TimelineShimmer from '@components/Shimmers/TimelineShimmer'
-import { Loader } from '@components/ui/Loader'
-import { NoDataFound } from '@components/ui/NoDataFound'
+import { Loader } from '@components/UIElements/Loader'
+import { NoDataFound } from '@components/UIElements/NoDataFound'
 import useAppStore from '@lib/store'
 import { LENSTUBE_VIDEOS_APP_ID } from '@utils/constants'
 import { FEED_QUERY } from '@utils/gql/queries'
