+++ src/components/Home/Timeline.tsx
@@ -1,4 +1,4 @@
-import VideoCard from '@components/common/VideoCard'
+import VideoCard from '@components/Common/VideoCard'
 import React, { FC } from 'react'
 import { LenstubePublication } from 'src/types/local'
