+++ src/components/Library/Sections/Recents.tsx
@@ -1,5 +1,5 @@
-import VideoCard from '@components/common/VideoCard'
-import { NoDataFound } from '@components/ui/NoDataFound'
+import VideoCard from '@components/Common/VideoCard'
+import { NoDataFound } from '@components/UIElements/NoDataFound'
 import useAppStore from '@lib/store'
 import { RECENTS_LIBRARY } from '@utils/url-path'
 import Link from 'next/link'
