+++ src/components/Library/SeeAll/Commented.tsx
@@ -1,10 +1,10 @@
 import { useQuery } from '@apollo/client'
-import Layout from '@components/common/Layout'
-import MetaTags from '@components/common/MetaTags'
+import Layout from '@components/Common/Layout'
+import MetaTags from '@components/Common/MetaTags'
 import Timeline from '@components/Home/Timeline'
 import TimelineShimmer from '@components/Shimmers/TimelineShimmer'
-import { Loader } from '@components/ui/Loader'
-import { NoDataFound } from '@components/ui/NoDataFound'
+import { Loader } from '@components/UIElements/Loader'
+import { NoDataFound } from '@components/UIElements/NoDataFound'
 import useAppStore from '@lib/store'
 import { LENSTUBE_VIDEOS_APP_ID } from '@utils/constants'
 import { PROFILE_FEED_QUERY } from '@utils/gql/queries'
