+++ src/components/Settings/index.tsx
@@ -1,5 +1,5 @@
-import Layout from '@components/common/Layout'
-import MetaTags from '@components/common/MetaTags'
+import Layout from '@components/Common/Layout'
+import MetaTags from '@components/Common/MetaTags'
 import React from 'react'
 
 const Settings = () => {
