+++ src/components/VideoDetails/MintVideo.tsx
@@ -1,8 +1,8 @@
 import { LENSHUB_PROXY_ABI } from '@abis/LensHubProxy'
 import { useMutation } from '@apollo/client'
-import { Button } from '@components/ui/Button'
-import { Loader } from '@components/ui/Loader'
-import Tooltip from '@components/ui/Tooltip'
+import { Button } from '@components/UIElements/Button'
+import { Loader } from '@components/UIElements/Loader'
+import Tooltip from '@components/UIElements/Tooltip'
 import { ERROR_MESSAGE, LENSHUB_PROXY_ADDRESS } from '@utils/constants'
 import omitKey from '@utils/functions/omitKey'
 import { CREATE_COLLECT_TYPED_DATA } from '@utils/gql/queries'
