+++ src/components/VideoDetails/SuggestedVideos.tsx
@@ -1,6 +1,6 @@
 import { useQuery } from '@apollo/client'
 import { SuggestedVideosShimmer } from '@components/Shimmers/VideoDetailShimmer'
-import { Loader } from '@components/ui/Loader'
+import { Loader } from '@components/UIElements/Loader'
 import { LENSTUBE_VIDEOS_APP_ID } from '@utils/constants'
 import getThumbnailUrl from '@utils/functions/getThumbnailUrl'
 import imageCdn from '@utils/functions/imageCdn'
