+++ src/components/VideoDetails/VideoComments.tsx
@@ -1,6 +1,6 @@
 import { useQuery } from '@apollo/client'
-import { Loader } from '@components/ui/Loader'
-import { NoDataFound } from '@components/ui/NoDataFound'
+import { Loader } from '@components/UIElements/Loader'
+import { NoDataFound } from '@components/UIElements/NoDataFound'
 import useAppStore from '@lib/store'
 import {
   LENSTUBE_COMMENTS_APP_ID,
