+++ src/components/common/ConnectWalletButton.tsx
@@ -1,6 +1,6 @@
-import { Button } from '@components/ui/Button'
-import { Loader } from '@components/ui/Loader'
-import Modal from '@components/ui/Modal'
+import { Button } from '@components/UIElements/Button'
+import { Loader } from '@components/UIElements/Loader'
+import Modal from '@components/UIElements/Modal'
 import useAppStore from '@lib/store'
 import { POLYGON_CHAIN_ID } from '@utils/constants'
 import { getWalletLogo } from '@utils/functions/getWalletLogo'
