+++ src/components/common/CreateChannel.tsx
@@ -1,7 +1,7 @@
 import { useMutation } from '@apollo/client'
-import { Button } from '@components/ui/Button'
-import { Input } from '@components/ui/Input'
-import Modal from '@components/ui/Modal'
+import { Button } from '@components/UIElements/Button'
+import { Input } from '@components/UIElements/Input'
+import Modal from '@components/UIElements/Modal'
 import useAppStore from '@lib/store'
 import { getHandle } from '@utils/functions/getHandle'
 import { getRandomProfilePicture } from '@utils/functions/getRandomProfilePicture'
