+++ src/components/common/Layout.tsx
@@ -1,7 +1,7 @@
-import CreateChannel from '@components/common/CreateChannel'
-import Header from '@components/common/Header'
-import MobileBottomNav from '@components/common/MobileBottomNav'
-import Sidebar from '@components/common/Sidebar'
+import CreateChannel from '@components/Common/CreateChannel'
+import Header from '@components/Common/Header'
+import MobileBottomNav from '@components/Common/MobileBottomNav'
+import Sidebar from '@components/Common/Sidebar'
 import useAppStore from '@lib/store'
 import { getToastOptions } from '@utils/functions/getToastOptions'
 import { useTheme } from 'next-themes'
