+++ src/pages/500.tsx
@@ -1,5 +1,5 @@
-import Layout from '@components/common/Layout'
-import { Button } from '@components/ui/Button'
+import Layout from '@components/Common/Layout'
+import { Button } from '@components/UIElements/Button'
 import { HOME } from '@utils/url-path'
 import Head from 'next/head'
 import Link from 'next/link'
